+++ server/src/log-records/entities/log-module-config.entity.ts
@@ -14,4 +14,7 @@ export class LogModuleConfig extends BaseEntity {
 
   @Column({ type: 'tinyint', default: false })
   enabled: boolean;
+
+  @Column({ type: 'simple-json', nullable: true })
+  enabledActions: string[] | null;
 }